明确了一些概念，增加了一些注释
