+++ spring-core/spring-context-default.go
@@ -67,479 +67,6 @@ func (item *beanCacheItem) store(t reflect.Type, bd *BeanDefinition, check bool)
 	return true
 }
 
-// wiringStack 存储绑定中的 Bean
-type wiringStack struct {
-	stack *list.List
-}
-
-// newWiringStack wiringStack 的构造函数
-func newWiringStack() *wiringStack {
-	return &wiringStack{
-		stack: list.New(),
-	}
-}
-
-// pushBack 添加一个 Item 到尾部
-func (s *wiringStack) pushBack(bd beanDefinition) {
-	s.stack.PushBack(bd)
-	SpringLogger.Tracef("wiring %s", bd.Description())
-}
-
-// popBack 删除尾部的 item
-func (s *wiringStack) popBack() {
-	e := s.stack.Remove(s.stack.Back())
-	SpringLogger.Tracef("wired %s", e.(beanDefinition).Description())
-}
-
-// path 返回依赖注入的路径
-func (s *wiringStack) path() (path string) {
-	for e := s.stack.Front(); e != nil; e = e.Next() {
-		w := e.Value.(beanDefinition)
-		path += fmt.Sprintf("=> %s ↩\n", w.Description())
-	}
-	return path[:len(path)-1]
-}
-
-// beanAssembly Bean 组装车间
-type beanAssembly interface {
-	springContext() SpringContext
-	collectBeans(v reflect.Value) bool
-	getBeanValue(v reflect.Value, beanId string, parent reflect.Value, field string) bool
-}
-
-// defaultBeanAssembly beanAssembly 的默认版本
-type defaultBeanAssembly struct {
-	springCtx   *defaultSpringContext
-	wiringStack *wiringStack
-}
-
-// newDefaultBeanAssembly defaultBeanAssembly 的构造函数
-func newDefaultBeanAssembly(springContext *defaultSpringContext) *defaultBeanAssembly {
-	return &defaultBeanAssembly{
-		springCtx:   springContext,
-		wiringStack: newWiringStack(),
-	}
-}
-
-func (beanAssembly *defaultBeanAssembly) springContext() SpringContext {
-	return beanAssembly.springCtx
-}
-
-// getCacheItem 获取指定类型的缓存项，返回值不会为 nil。
-func (beanAssembly *defaultBeanAssembly) getCacheItem(t reflect.Type) *beanCacheItem {
-	beanCache := &beanAssembly.springCtx.beanCacheByType
-
-	// 如果缓存已存在则直接返回
-	if c, ok := beanCache.Load(t); ok {
-		return c.(*beanCacheItem)
-	}
-
-	result := newBeanCacheItem()
-	beanCache.Store(t, result)
-	return result
-}
-
-// getBeanValue 根据 BeanId 查找 Bean 并返回 Bean 源的值
-func (beanAssembly *defaultBeanAssembly) getBeanValue(v reflect.Value, beanId string, parent reflect.Value, field string) bool {
-
-	var (
-		ok       bool
-		beanType reflect.Type
-	)
-
-	if beanType, ok = ValidBeanValue(v); !ok {
-		panic(fmt.Errorf("receiver must be ref type, bean: \"%s\" field: %s", beanId, field))
-	}
-
-	result := make([]*BeanDefinition, 0)
-
-	typeName, beanName, nullable := ParseBeanId(beanId)
-	m := beanAssembly.getCacheItem(beanType)
-	for _, bean := range m.beans {
-		// 不能将自身赋给自身的字段 && 类型全限定名匹配
-		if bean.Value() != parent && bean.Match(typeName, beanName) {
-			result = append(result, bean)
-		}
-	}
-
-	// 对接口匹配开个绿灯，如果指定了 Bean 名称则尝试通过名称获取以防没有通过 Export 显示导出接口
-	if beanType.Kind() == reflect.Interface && beanName != "" {
-		beanCache := beanAssembly.springCtx.beanCacheByName
-		if cache, o := beanCache[beanName]; o {
-			for _, b := range cache {
-				// 不能将自身赋给自身的字段 && 类型匹配
-				if b.Value() != parent && b.Type().AssignableTo(beanType) && b.Match(typeName, beanName) {
-					found := false // 排重
-					for _, r := range result {
-						if r == b {
-							found = true
-							break
-						}
-					}
-					if !found {
-						result = append(result, b)
-						SpringLogger.Warnf("you should call Export() on %s", b.Description())
-					}
-				}
-			}
-		}
-	}
-
-	count := len(result)
-
-	// 没有找到
-	if count == 0 {
-		if nullable {
-			return false
-		} else {
-			panic(fmt.Errorf("can't find bean, bean: \"%s\" field: %s type: %s", beanId, field, beanType))
-		}
-	}
-
-	var primaryBeans []*BeanDefinition
-
-	for _, bean := range result {
-		if bean.primary {
-			primaryBeans = append(primaryBeans, bean)
-		}
-	}
-
-	if len(primaryBeans) > 1 {
-		msg := fmt.Sprintf("found %d primary beans, bean: \"%s\" field: %s type: %s [", len(primaryBeans), beanId, field, beanType)
-		for _, b := range primaryBeans {
-			msg += "( " + b.Description() + " ), "
-		}
-		msg = msg[:len(msg)-2] + "]"
-		panic(errors.New(msg))
-	}
-
-	if len(primaryBeans) == 0 {
-		if count > 1 {
-			msg := fmt.Sprintf("found %d beans, bean: \"%s\" field: %s type: %s [", len(result), beanId, field, beanType)
-			for _, b := range result {
-				msg += "( " + b.Description() + " ), "
-			}
-			msg = msg[:len(msg)-2] + "]"
-			panic(errors.New(msg))
-		}
-		primaryBeans = append(primaryBeans, result[0])
-	}
-
-	// 依赖注入
-	beanAssembly.wireBeanDefinition(primaryBeans[0], false)
-
-	// 恰好 1 个
-	v0 := SpringUtils.ValuePatchIf(v, beanAssembly.springCtx.AllAccess())
-	v0.Set(primaryBeans[0].Value())
-	return true
-}
-
-// collectBeans 收集符合条件的 Bean 源
-func (beanAssembly *defaultBeanAssembly) collectBeans(v reflect.Value) bool {
-
-	t := v.Type()
-	ev := reflect.New(t).Elem()
-
-	// 查找数组类型
-	{
-		m := beanAssembly.getCacheItem(t)
-		for _, d := range m.beans {
-			for i := 0; i < d.Value().Len(); i++ {
-				di := d.Value().Index(i)
-				if di.Kind() == reflect.Struct {
-					beanAssembly.wireSliceItem(di.Addr(), d)
-				} else if di.Kind() == reflect.Ptr {
-					if de := di.Elem(); de.Kind() == reflect.Struct {
-						beanAssembly.wireSliceItem(di, d)
-					}
-				}
-				ev = reflect.Append(ev, di)
-			}
-		}
-	}
-
-	// 查找单例类型
-	{
-		if et := t.Elem(); IsRefType(et.Kind()) {
-			m := beanAssembly.getCacheItem(et)
-			for _, d := range m.beans {
-				beanAssembly.wireBeanDefinition(d, false)
-				ev = reflect.Append(ev, d.Value())
-			}
-		}
-	}
-
-	if ev.Len() > 0 {
-		v = SpringUtils.ValuePatchIf(v, beanAssembly.springCtx.AllAccess())
-		v.Set(ev)
-		return true
-	}
-	return false
-}
-
-// fieldBeanDefinition 带字段名称的 BeanDefinition 实现
-type fieldBeanDefinition struct {
-	*BeanDefinition
-	field string // 字段名称
-}
-
-// Description 返回 Bean 的详细描述
-func (d *fieldBeanDefinition) Description() string {
-	return fmt.Sprintf("%s field: %s %s", d.bean.beanClass(), d.field, d.FileLine())
-}
-
-// delegateBeanDefinition 代理功能的 BeanDefinition 实现
-type delegateBeanDefinition struct {
-	*BeanDefinition
-	delegate beanDefinition // 代理项
-}
-
-// Description 返回 Bean 的详细描述
-func (d *delegateBeanDefinition) Description() string {
-	return fmt.Sprintf("%s value %s", d.delegate.springBean().beanClass(), d.delegate.FileLine())
-}
-
-// wireSliceItem 注入 slice 的元素值
-func (beanAssembly *defaultBeanAssembly) wireSliceItem(v reflect.Value, d beanDefinition) {
-	bd := ValueToBeanDefinition("", v)
-	bd.file = d.getFile()
-	bd.line = d.getLine()
-	beanAssembly.wireBeanDefinition(bd, false)
-}
-
-// wireBeanDefinition 绑定 BeanDefinition 指定的 Bean
-func (beanAssembly *defaultBeanAssembly) wireBeanDefinition(bd beanDefinition, onlyAutoWire bool) {
-
-	// 是否已删除
-	if bd.getStatus() == beanStatus_Deleted {
-		panic(fmt.Errorf("bean: \"%s\" have been deleted", bd.BeanId()))
-	}
-
-	// 是否已绑定
-	if bd.getStatus() == beanStatus_Wired {
-		return
-	}
-
-	// 将当前 Bean 放入注入栈，以便检测循环依赖。
-	beanAssembly.wiringStack.pushBack(bd)
-
-	// 是否循环依赖
-	if bd.getStatus() == beanStatus_Wiring {
-		if _, ok := bd.springBean().(*objectBean); !ok {
-			panic(errors.New("found circle autowire"))
-		}
-		return
-	}
-
-	bd.setStatus(beanStatus_Wiring)
-
-	// 首先初始化当前 Bean 不直接依赖的那些 Bean
-	for _, selector := range bd.getDependsOn() {
-		if bean, ok := beanAssembly.springCtx.FindBean(selector); !ok {
-			panic(fmt.Errorf("can't find bean: \"%v\"", selector))
-		} else {
-			beanAssembly.wireBeanDefinition(bean, false)
-		}
-	}
-
-	// 如果是成员方法 Bean，需要首先初始化它的父 Bean
-	if mBean, ok := bd.springBean().(*methodBean); ok {
-		beanAssembly.wireBeanDefinition(mBean.parent, false)
-	}
-
-	switch bean := bd.springBean().(type) {
-	case *objectBean: // 原始对象
-		beanAssembly.wireObjectBean(bd, onlyAutoWire)
-	case *constructorBean: // 构造函数
-		fnValue := reflect.ValueOf(bean.fn)
-		beanAssembly.wireFunctionBean(fnValue, &bean.functionBean, bd)
-	case *methodBean: // 成员方法
-		fnValue := bean.parent.Value().MethodByName(bean.method)
-		beanAssembly.wireFunctionBean(fnValue, &bean.functionBean, bd)
-	default:
-		panic(errors.New("unknown spring bean type"))
-	}
-
-	// 如果有则执行用户设置的初始化函数
-	if init := bd.getInit(); init != nil {
-		if err := init.run(beanAssembly.springCtx); err != nil {
-			panic(err)
-		}
-	}
-
-	bd.setStatus(beanStatus_Wired)
-
-	// 删除保存的注入帧
-	beanAssembly.wiringStack.popBack()
-}
-
-// wireObjectBean 对原始对象进行注入
-func (beanAssembly *defaultBeanAssembly) wireObjectBean(bd beanDefinition, onlyAutoWire bool) {
-	st := bd.Type()
-	switch sk := st.Kind(); sk {
-	case reflect.Slice:
-		et := st.Elem()
-		if ek := et.Kind(); ek == reflect.Struct { // 结构体数组
-			v := bd.Value()
-			for i := 0; i < v.Len(); i++ {
-				iv := v.Index(i).Addr()
-				beanAssembly.wireSliceItem(iv, bd)
-			}
-		} else if ek == reflect.Ptr { // 指针数组
-			it := et.Elem()
-			if ik := it.Kind(); ik == reflect.Struct { // 结构体指针数组
-				v := bd.Value()
-				for p := 0; p < v.Len(); p++ {
-					pv := v.Index(p)
-					beanAssembly.wireSliceItem(pv, bd)
-				}
-			}
-		}
-	case reflect.Ptr:
-		if et := st.Elem(); et.Kind() == reflect.Struct { // 结构体指针
-
-			var etName string
-			if etName = et.Name(); etName == "" {
-				etName = et.String()
-			}
-
-			sv := bd.Value()
-			ev := sv.Elem()
-
-			for i := 0; i < et.NumField(); i++ {
-				// 字段包含 value 标签时嵌套处理只注入变量
-				fieldOnlyAutoWire := false
-
-				ft := et.Field(i)
-				fv := ev.Field(i)
-
-				fieldName := etName + ".$" + ft.Name
-
-				// 处理 value 标签
-				if !onlyAutoWire {
-					// 避免父结构体有 value 标签重新解析导致失败的情况
-					if tag, ok := ft.Tag.Lookup("value"); ok {
-						fieldOnlyAutoWire = true
-						bindStructField(beanAssembly.springCtx, fv, tag, bindOption{
-							fieldName: fieldName,
-							allAccess: beanAssembly.springCtx.AllAccess(),
-						})
-					}
-				}
-
-				// 处理 autowire 标签
-				if beanId, ok := ft.Tag.Lookup("autowire"); ok {
-					beanAssembly.wireStructField(fv, beanId, sv, fieldName)
-				}
-
-				// 处理 inject 标签
-				if beanId, ok := ft.Tag.Lookup("inject"); ok {
-					beanAssembly.wireStructField(fv, beanId, sv, fieldName)
-				}
-
-				// 处理结构体类型的字段，防止递归所以不支持指针结构体字段
-				if ft.Type.Kind() == reflect.Struct {
-					// 开放私有字段，但是不会更新原属性
-					fv0 := SpringUtils.ValuePatchIf(fv, beanAssembly.springCtx.AllAccess())
-					if fv0.CanSet() {
-
-						b := ValueToBeanDefinition("", fv0.Addr())
-						b.file = bd.getFile()
-						b.line = bd.getLine()
-						fbd := &fieldBeanDefinition{b, fieldName}
-						beanAssembly.wireBeanDefinition(fbd, fieldOnlyAutoWire)
-					}
-				}
-			}
-		}
-	}
-}
-
-// wireFunctionBean 对函数定义 Bean 进行注入
-func (beanAssembly *defaultBeanAssembly) wireFunctionBean(fnValue reflect.Value, bean *functionBean, bd beanDefinition) {
-
-	var in []reflect.Value
-
-	if bean.stringArg != nil {
-		if r := bean.stringArg.Get(beanAssembly, bd); len(r) > 0 {
-			in = append(in, r...)
-		}
-	}
-
-	if bean.optionArg != nil {
-		if r := bean.optionArg.Get(beanAssembly, bd); len(r) > 0 {
-			in = append(in, r...)
-		}
-	}
-
-	out := fnValue.Call(in)
-
-	// 获取第一个返回值
-	val := out[0]
-
-	// 检查是否有 error 返回
-	if len(out) == 2 {
-		if err := out[1].Interface(); err != nil {
-			panic(fmt.Errorf("function bean: \"%s\" return error: %v", bd.FileLine(), err))
-		}
-	}
-
-	if IsRefType(val.Kind()) {
-		// 如果实现接口的是值类型，那么需要转换成指针类型然后赋给接口
-		if val.Kind() == reflect.Interface && IsValueType(val.Elem().Kind()) {
-			ptrVal := reflect.New(val.Elem().Type())
-			ptrVal.Elem().Set(val.Elem())
-			bean.rValue.Set(ptrVal)
-		} else {
-			bean.rValue.Set(val)
-		}
-	} else {
-		bean.rValue.Elem().Set(val)
-	}
-
-	if bean.Value().IsNil() {
-		panic(fmt.Errorf("function bean: \"%s\" return nil", bd.FileLine()))
-	}
-
-	// 对返回值进行依赖注入
-	b := &BeanDefinition{
-		name:   bd.Name(),
-		status: beanStatus_Default,
-		file:   bd.getFile(),
-		line:   bd.getLine(),
-	}
-
-	if bean.Type().Kind() == reflect.Interface {
-		b.bean = newObjectBean(bean.Value().Elem())
-	} else {
-		b.bean = newObjectBean(bean.Value())
-	}
-
-	beanAssembly.wireBeanDefinition(&delegateBeanDefinition{b, bd}, false)
-}
-
-// wireStructField 对结构体的字段进行绑定
-func (beanAssembly *defaultBeanAssembly) wireStructField(v reflect.Value,
-	beanId string, parent reflect.Value, field string) {
-
-	_, beanName, nullable := ParseBeanId(beanId)
-	if beanName == "[]" { // 收集模式，autowire:"[]"
-
-		// 收集模式的绑定对象必须是数组
-		if v.Type().Kind() != reflect.Slice {
-			panic(fmt.Errorf("field: %s should be slice", field))
-		}
-
-		ok := beanAssembly.collectBeans(v)
-		if !ok && !nullable { // 没找到且不能为空则 panic
-			panic(fmt.Errorf("can't find bean: \"%s\" field: %s", beanId, field))
-		}
-
-	} else { // 匹配模式，autowire:"" or autowire:"name"
-		beanAssembly.getBeanValue(v, beanId, parent, field)
-	}
-}
-
 // defaultSpringContext SpringContext 的默认版本
 type defaultSpringContext struct {
 	// 属性值列表接口
@@ -665,7 +192,7 @@ func (ctx *defaultSpringContext) RegisterNameBeanFn(name string, fn interface{},
 // selector 可以是 *BeanDefinition，可以是 BeanId，还可以是 (Type)(nil) 变量。
 // 必须给定方法名而不能通过遍历方法列表比较方法类型的方式获得函数名，因为不同方法的类型可能相同。
 // 而且 interface 的方法类型不带 receiver 而成员方法的类型带有 receiver，两者类型不好匹配。
-func (ctx *defaultSpringContext) RegisterMethodBean(selector interface{}, method string, tags ...string) *BeanDefinition {
+func (ctx *defaultSpringContext) RegisterMethodBean(selector BeanSelector, method string, tags ...string) *BeanDefinition {
 	return ctx.RegisterNameMethodBean("", selector, method, tags...)
 }
 
@@ -673,7 +200,7 @@ func (ctx *defaultSpringContext) RegisterMethodBean(selector interface{}, method
 // selector 可以是 *BeanDefinition，可以是 BeanId，还可以是 (Type)(nil) 变量。
 // 必须给定方法名而不能通过遍历方法列表比较方法类型的方式获得函数名，因为不同方法的类型可能相同。
 // 而且 interface 的方法类型不带 receiver 而成员方法的类型带有 receiver，两者类型不好匹配。
-func (ctx *defaultSpringContext) RegisterNameMethodBean(name string, selector interface{}, method string, tags ...string) *BeanDefinition {
+func (ctx *defaultSpringContext) RegisterNameMethodBean(name string, selector BeanSelector, method string, tags ...string) *BeanDefinition {
 	ctx.checkRegistration()
 
 	if selector == nil || selector == "" {
@@ -740,10 +267,15 @@ func (ctx *defaultSpringContext) GetBeanByName(beanId string, i interface{}) boo
 
 // FindBean 获取单例 Bean，若多于 1 个则 panic；找到返回 true 否则返回 false。
 // selector 可以是 BeanId，还可以是 (Type)(nil) 变量，Type 为接口类型时带指针。
-func (ctx *defaultSpringContext) FindBean(selector interface{}) (*BeanDefinition, bool) {
+func (ctx *defaultSpringContext) FindBean(selector BeanSelector) (*BeanDefinition, bool) {
+	return ctx.FindBeanByName(GetBeanId(selector))
+}
+
+// FindBeanByName 根据名称和类型获取单例 Bean，若多于 1 个则 panic；找到返回 true 否则返回 false。
+func (ctx *defaultSpringContext) FindBeanByName(beanId string) (*BeanDefinition, bool) {
 	ctx.checkAutoWired()
 
-	typeName, beanName, _ := ParseBeanId(BeanId(selector))
+	typeName, beanName, _ := ParseBeanId(beanId)
 
 	result := make([]*BeanDefinition, 0)
 	for _, bean := range ctx.beanMap {
@@ -772,7 +304,7 @@ func (ctx *defaultSpringContext) FindBean(selector interface{}) (*BeanDefinition
 
 	// 多于 1 个
 	if count > 1 {
-		msg := fmt.Sprintf("found %d beans, bean: \"%v\" [", len(result), selector)
+		msg := fmt.Sprintf("found %d beans, bean: \"%v\" [", len(result), beanId)
 		for _, b := range result {
 			msg += "( " + b.Description() + " ), "
 		}
@@ -782,11 +314,6 @@ func (ctx *defaultSpringContext) FindBean(selector interface{}) (*BeanDefinition
 
 	// 恰好 1 个 & 仅供查询无需绑定
 	return result[0], true
-}
-
-// FindBeanByName 根据名称和类型获取单例 Bean，若多于 1 个则 panic；找到返回 true 否则返回 false。
-func (ctx *defaultSpringContext) FindBeanByName(beanId string) (*BeanDefinition, bool) {
-	return ctx.FindBean(beanId)
 }
 
 // CollectBeans 收集数组或指针定义的所有符合条件的 Bean 对象，收集到返回 true，否则返回 false。
@@ -916,56 +443,25 @@ func (ctx *defaultSpringContext) resolveBean(bd *BeanDefinition) {
 
 // registerMethodBeans 注册方法 Bean
 func (ctx *defaultSpringContext) registerMethodBeans() {
-	var (
-		selector string
-		filter   func(*BeanDefinition) bool
-	)
 	for _, bd := range ctx.methodBeans {
 		bean := bd.bean.(*fakeMethodBean)
 		result := make([]*BeanDefinition, 0)
 
-		switch e := bean.selector.(type) {
-		case *BeanDefinition:
-			selector = e.BeanId()
-			result = append(result, e)
-		case string:
-			selector = e
-			typeName, beanName, _ := ParseBeanId(e)
-			filter = func(b *BeanDefinition) bool {
-				return b.Match(typeName, beanName)
-			}
-		case reflect.Type:
-			selector = e.String()
-			filter = func(b *BeanDefinition) bool {
-				return b.Type() == e
-			}
-		default:
-			t := reflect.TypeOf(e)
-			// 如果是接口类型需要解除外层指针
-			if t.Kind() == reflect.Ptr && t.Elem().Kind() == reflect.Interface {
-				t = t.Elem()
-			}
-			selector = t.String()
-			filter = func(b *BeanDefinition) bool {
-				return b.Type() == t
-			}
-		}
+		typeName, beanName, _ := ParseBeanId(bean.parent)
 
-		if filter != nil {
-			for _, b := range ctx.beanMap {
-				if filter(b) {
-					result = append(result, b)
-				}
+		for _, b := range ctx.beanMap {
+			if b.Match(typeName, beanName) {
+				result = append(result, b)
 			}
 		}
 
 		if l := len(result); l == 0 {
-			panic(fmt.Errorf("can't find parent bean: \"%s\"", selector))
+			panic(fmt.Errorf("can't find parent bean: \"%s\"", bean.parent))
 		} else if l > 1 {
-			panic(fmt.Errorf("found %d parent bean: \"%s\"", l, selector))
+			panic(fmt.Errorf("found %d parent bean: \"%s\"", l, bean.parent))
 		}
 
-		bd.bean = newMethodBean(result[0], bean.method, bean.tags...)
+		bd.bean = newMethodBean(result[0], bean.method, bean.tags)
 		if bd.name == "" { // 使用默认名称
 			bd.name = bd.bean.Type().String()
 		}
