+++ spring-core/spring-context.go
@@ -80,13 +80,13 @@ type SpringContext interface {
 	// selector 可以是 *BeanDefinition，可以是 BeanId，还可以是 (Type)(nil) 变量。
 	// 必须给定方法名而不能通过遍历方法列表比较方法类型的方式获得函数名，因为不同方法的类型可能相同。
 	// 而且 interface 的方法类型不带 receiver 而成员方法的类型带有 receiver，两者类型不好匹配。
-	RegisterMethodBean(selector interface{}, method string, tags ...string) *BeanDefinition
+	RegisterMethodBean(selector BeanSelector, method string, tags ...string) *BeanDefinition
 
 	// RegisterNameMethodBean 注册成员方法单例 Bean，需指定名称，重复注册会 panic。
 	// selector 可以是 *BeanDefinition，可以是 BeanId，还可以是 (Type)(nil) 变量。
 	// 必须给定方法名而不能通过遍历方法列表比较方法类型的方式获得函数名，因为不同方法的类型可能相同。
 	// 而且 interface 的方法类型不带 receiver 而成员方法的类型带有 receiver，两者类型不好匹配。
-	RegisterNameMethodBean(name string, selector interface{}, method string, tags ...string) *BeanDefinition
+	RegisterNameMethodBean(name string, selector BeanSelector, method string, tags ...string) *BeanDefinition
 
 	// @Incubate 注册成员方法单例 Bean，不指定名称，重复注册会 panic。
 	RegisterMethodBeanFn(method interface{}, tags ...string) *BeanDefinition
@@ -112,7 +112,7 @@ type SpringContext interface {
 
 	// FindBean 获取单例 Bean，若多于 1 个则 panic；找到返回 true 否则返回 false。
 	// selector 可以是 BeanId，还可以是 (*Type)(nil) 变量，Type 为接口类型时带指针。
-	FindBean(selector interface{}) (*BeanDefinition, bool)
+	FindBean(selector BeanSelector) (*BeanDefinition, bool)
 
 	// FindBeanByName 根据名称和类型获取单例 Bean，若多于 1 个则 panic；找到返回 true 否则返回 false。
 	FindBeanByName(beanId string) (*BeanDefinition, bool)
