--- spring-core/spring-utils.go
@@ -1,38 +1,0 @@
-/*
- * Copyright 2012-2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package SpringCore
-
-import (
-	"fmt"
-	"reflect"
-)
-
-type FileAddr interface {
-	FileLine() string
-}
-
-type FileAddress struct {
-	File string // 文件路径
-	Line int    // 行号
-}
-
-func (f *FileAddress) FileLine() string {
-	return fmt.Sprintf("%s:%d", f.File, f.Line)
-}
-
-// ErrorType error 的反射类型
-var ErrorType = reflect.TypeOf((*error)(nil)).Elem()
